% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_effects_influence.r
\name{calculate_effects_influence}
\alias{calculate_effects_influence}
\title{Calculate Effects-based Influence}
\usage{
calculate_effects_influence(x, term_info)
}
\arguments{
\item{x}{gam_influence object}

\item{term_info}{Information about the term}
}
\value{
Numeric influence score
}
\description{
Calculate Effects-based Influence
}
\keyword{internal}
