% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fit_statistics.R
\name{calculate_fit_statistics}
\alias{calculate_fit_statistics}
\title{Calculate Fit Statistics}
\usage{
calculate_fit_statistics(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
List with fit statistics
}
\description{
Calculate Fit Statistics
}
\keyword{internal}
