% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gam_influence.r
\name{calculate_gam_influence}
\alias{calculate_gam_influence}
\title{Calculate Influence Statistics}
\usage{
calculate_gam_influence(x, ...)
}
\arguments{
\item{x}{A gam_influence object}

\item{...}{Additional arguments}
}
\value{
The gam_influence object with calculated statistics
}
\description{
Perform influence calculations for the GAM model
}
\keyword{internal}
