% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_influence.R
\name{calculate_influence}
\alias{calculate_influence}
\title{Perform Influence Calculations}
\usage{
calculate_influence(
  obj,
  islog = NULL,
  rescale_method = c("geometric_mean", "arithmetic_mean", "raw", "custom"),
  custom_rescale_value = 1,
  confidence_level = 0.95,
  family_method = c("auto", "gaussian", "binomial", "gamma", "poisson"),
  subset_var = NULL,
  subset_value = NULL,
  ...
)
}
\arguments{
\item{obj}{An object for which to perform calculations.}

\item{...}{Additional arguments passed to methods.}
}
\description{
A generic S3 function to perform calculations on a `gam_influence` object.
}
