% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model_progression.R
\name{calculate_model_progression}
\alias{calculate_model_progression}
\title{Calculate Model Progression Statistics}
\usage{
calculate_model_progression(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
Data frame with progression statistics
}
\description{
Build models progressively and calculate fit statistics
}
\keyword{internal}
