% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_parametric_focus_effects.R
\name{calculate_parametric_effects}
\alias{calculate_parametric_effects}
\title{calculate_parametric_effects}
\usage{
calculate_parametric_effects(model, focus_term, data)
}
\arguments{
\item{model}{gam object}

\item{focus_term}{Focus term}

\item{data}{Data used for the model}
}
\value{
List with influence metrics per focus level
}
\description{
Calculate influence of a term for each level of the focus term,
}
\keyword{internal}
