% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_smooth_effects.r
\name{calculate_smooth_effects}
\alias{calculate_smooth_effects}
\title{Calculate Smooth Term Effects}
\usage{
calculate_smooth_effects(model, term, data)
}
\arguments{
\item{model}{GAM model}

\item{term}{Smooth term label}

\item{data}{Data}
}
\value{
List with smooth effects
}
\description{
Calculate Smooth Term Effects
}
\keyword{internal}
