% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_smooth_influences.R
\name{calculate_smooth_influences}
\alias{calculate_smooth_influences}
\title{Calculate Smooth Term Influences on Parametric Focus Term}
\usage{
calculate_smooth_influences(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
List with influence calculations per focus level
}
\description{
Calculate Smooth Term Influences on Parametric Focus Term
}
\keyword{internal}
