% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_step_effects.R
\name{calculate_step_effects}
\alias{calculate_step_effects}
\title{Calculate Focus Term Effects at Each Step}
\usage{
calculate_step_effects(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
List with effects at each step
}
\description{
Calculate Focus Term Effects at Each Step
}
\keyword{internal}
