% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_term_effects.r
\name{calculate_term_effects}
\alias{calculate_term_effects}
\title{Calculate Effects for a Specific Term}
\usage{
calculate_term_effects(model, term, data)
}
\arguments{
\item{model}{GAM model}

\item{term}{Term name}

\item{data}{Data used for fitting}
}
\value{
List with term effects
}
\description{
Calculate Effects for a Specific Term
}
\keyword{internal}
