% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_term_influences.R
\name{calculate_term_influences}
\alias{calculate_term_influences}
\title{Calculate Influence of Terms}
\usage{
calculate_term_influences(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
List with influence calculations
}
\description{
Calculate Influence of Terms
}
\keyword{internal}
