% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_influence_cdi_plots.r
\name{cdi_plot_1d}
\alias{cdi_plot_1d}
\title{CDI plotting methods for GAMInfluence class}
\usage{
cdi_plot_1d(term, variable, colour, n_bins, smooth_info)
}
\arguments{
\item{term}{The term to plot}

\item{variable}{The variable to plot (optional)}

\item{colour}{The colour to use for the plot}

\item{n_bins}{The number of bins to use for the plot}

\item{smooth_info}{Information about the smooth term}
}
\value{
A list of ggplot objects for the CDI plots
}
\description{
# CDI plot for 1D smooths
}
\keyword{internal}
