% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdi.R
\name{create_model_summary_table}
\alias{create_model_summary_table}
\title{Create Model Summary Table for CDI Plot}
\usage{
create_model_summary_table(obj)
}
\arguments{
\item{obj}{A gam_influence object}
}
\value{
A ggplot object containing the summary table
}
\description{
Creates a table showing model deviance and R² information
}
