% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_by_variables.r
\name{expand_by_variables}
\alias{expand_by_variables}
\title{Expand By-Variables into Separate Terms}
\usage{
expand_by_variables(model)
}
\arguments{
\item{model}{GAM model}
}
\value{
List with expanded term information
}
\description{
Expand By-Variables into Separate Terms
}
\keyword{internal}
