% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Updated export_results function to include per-level influences}
\usage{
export_results(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
Data frame with influence results including per-level data
}
\description{
Updated export_results function to include per-level influences
}
