% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clean_levels.R
\name{extract_clean_levels}
\alias{extract_clean_levels}
\title{Create clean labels for terms}
\usage{
extract_clean_levels(coef_names, focus_term)
}
\arguments{
\item{coef_names}{Coefficient names from the GAM model}

\item{focus_term}{Focus term name}
}
\value{
Vector of cleaned up names
}
\description{
Create clean labels for terms
}
\keyword{internal}
