% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_focus_effects.R
\name{extract_focus_effects}
\alias{extract_focus_effects}
\title{Extract Focus Term Effects from Model}
\usage{
extract_focus_effects(model, focus_term)
}
\arguments{
\item{model}{GAM model}

\item{focus_term}{Focus term name}
}
\value{
Vector of relative effects
}
\description{
Extract Focus Term Effects from Model
}
\keyword{internal}
