% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_indices.R
\name{extract_indices.default}
\alias{extract_indices.default}
\title{Default Method for extract_indices}
\usage{
\method{extract_indices}{default}(obj)
}
\arguments{
\item{obj}{An object of unsupported class}
}
\value{
Throws an error
}
\description{
Default S3 method that provides an informative error message
}
