% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_indices.R
\name{extract_indices.gam_influence}
\alias{extract_indices.gam_influence}
\title{Extract Indices from gam_influence Object}
\usage{
\method{extract_indices}{gam_influence}(obj)
}
\arguments{
\item{obj}{A \code{gam_influence} object containing calculated indices from \code{calculate_influence()}.}
}
\value{
A data frame with standardised and unstandardised indices
}
\description{
S3 method for extracting indices from gam_influence objects
}
