% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{gam_influence_from_box_cox}
\alias{gam_influence_from_box_cox}
\title{Convert a Box-Cox Transformed GAM to gam_influence object}
\usage{
gam_influence_from_box_cox(bc_model, focus, ...)
}
\arguments{
\item{bc_model}{An object of class "gam_box_cox" from box_cox_gam().}

\item{focus}{Character. Name of the focus term for influence analysis.}

\item{...}{Additional arguments passed to gam_influence.}
}
\value{
A gam_influence object.
}
\description{
Create a gam_influence object from a Box-Cox transformed GAM for use with
  influence diagnostics and other gamInflu functions.
}
