% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_summary.r
\name{get_model_summary}
\alias{get_model_summary}
\title{Get Model Summary}
\usage{
get_model_summary(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
List with model summary information
}
\description{
Get Model Summary
}
