% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_per_level_influences.R
\name{get_per_level_influences}
\alias{get_per_level_influences}
\title{Get Per-Level Influences for a Specific Term}
\usage{
get_per_level_influences(x, term)
}
\arguments{
\item{x}{gam_influence object}

\item{term}{Term name}
}
\value{
Named vector of influences per focus level
}
\description{
Extract the per-level influence values for detailed analysis
}
