% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r2.r
\name{get_r2}
\alias{get_r2}
\title{Get Model r2}
\usage{
get_r2(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
The model_progression object from the calculated gam_influence object
}
\description{
Get Model r2
}
