% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma_family.r
\name{ggamma-distribution}
\alias{ggamma-distribution}
\alias{dgg}
\alias{dgg.default}
\alias{pgg}
\alias{pgg.default}
\alias{qgg}
\alias{qgg.default}
\alias{rgg}
\alias{rgg.default}
\title{Generalized Gamma Distribution Functions}
\usage{
dgg(x, mu, sigma, Q, log = FALSE)

\method{dgg}{default}(x, mu, sigma, Q, log = FALSE)

pgg(q, mu, sigma, Q, lower.tail = TRUE, log.p = FALSE)

\method{pgg}{default}(q, mu, sigma, Q, lower.tail = TRUE, log.p = FALSE)

qgg(p, mu, sigma, Q, lower.tail = TRUE, log.p = FALSE)

\method{qgg}{default}(p, mu, sigma, Q, lower.tail = TRUE, log.p = FALSE)

rgg(n, mu, sigma, Q)

\method{rgg}{default}(n, mu, sigma, Q)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter (log scale)}

\item{sigma}{scale parameter (positive)}

\item{Q}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X \\le x], otherwise P[X > x]}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
S3 methods for the 3-parameter generalized gamma distribution,
compatible with mgcv's extended family framework.
}
\details{
This implements a generalized gamma distribution with three parameters:
\itemize{
  \item \code{mu}: location parameter (log scale)
  \item \code{sigma}: scale parameter (positive)
  \item \code{Q}: shape parameter
}

The generalized gamma reduces to special cases:
\itemize{
  \item When Q \\to 0: Log-normal distribution
  \item When Q = 1: Weibull distribution
  \item When sigma = 1: Gamma distribution
}
}
