% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_variable_for_cdi.R
\name{infer_variable_for_cdi}
\alias{infer_variable_for_cdi}
\title{Infer Variable Name for CDI Plot}
\usage{
infer_variable_for_cdi(x, term_info)
}
\arguments{
\item{x}{gam_influence object}

\item{term_info}{Term information}
}
\value{
Variable name
}
\description{
Infer Variable Name for CDI Plot
}
\keyword{internal}
