% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_focus_related.R
\name{is_focus_related}
\alias{is_focus_related}
\title{Check if term is related to the focus term}
\usage{
is_focus_related(term_name, focus_term)
}
\arguments{
\item{term_name}{Name of the term to check}

\item{focus_term}{Name of the focus term}
}
\value{
Logical indicating if term is focus-related
}
\description{
Check if term is related to the focus term
}
\keyword{internal}
