% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_parametric_term.r
\name{is_parametric_term}
\alias{is_parametric_term}
\title{Check if Term is Parametric}
\usage{
is_parametric_term(model, term)
}
\arguments{
\item{model}{GAM model}

\item{term}{Term name to check}
}
\value{
Logical indicating if term is parametric
}
\description{
Check if Term is Parametric
}
\keyword{internal}
