% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tensor_product.r
\name{is_tensor_product}
\alias{is_tensor_product}
\title{Check if Term is Tensor Product}
\usage{
is_tensor_product(term_label)
}
\arguments{
\item{term_label}{Term label}
}
\value{
Logical
}
\description{
Check if Term is Tensor Product
}
\keyword{internal}
