% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdi_coefficients.R
\name{plot_cdi_coefficients}
\alias{plot_cdi_coefficients}
\title{Plot CDI Coefficients}
\usage{
plot_cdi_coefficients(x, term, ...)
}
\arguments{
\item{x}{A gam_influence object}

\item{term}{The term to plot}
}
\value{
A ggplot object for the CDI plot
}
\description{
Plot CDI Coefficients
}
\keyword{internal}
