% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdi_distribution.R
\name{plot_cdi_distribution}
\alias{plot_cdi_distribution}
\title{Plot CDI Distribution Panel (ggplot)}
\usage{
plot_cdi_distribution(x, variable, ...)
}
\arguments{
\item{x}{gam_influence object}

\item{variable}{Variable name for distribution}

\item{...}{Additional plotting arguments}
}
\value{
ggplot object
}
\description{
Plot CDI Distribution Panel (ggplot)
}
\keyword{internal}
