% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdi_influence.R
\name{plot_cdi_influence}
\alias{plot_cdi_influence}
\title{Plot CDI Influence Panel (ggplot)}
\usage{
plot_cdi_influence(x, term, ...)
}
\arguments{
\item{x}{gam_influence object}

\item{term}{Term name}

\item{...}{Additional plotting arguments}
}
\value{
ggplot object
}
\description{
Plot CDI Influence Panel (ggplot)
}
\keyword{internal}
