% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_numeric.R
\name{plot_distribution_numeric}
\alias{plot_distribution_numeric}
\title{Plot Distribution for Numeric Smooth Variable (ggplot)}
\usage{
plot_distribution_numeric(focus_data, smooth_data, focus_name, variable_name)
}
\arguments{
\item{focus_data}{Focus term data}

\item{smooth_data}{Smooth variable data}

\item{focus_name}{Focus term name}

\item{variable_name}{Variable name}
}
\value{
ggplot object
}
\description{
Plot Distribution for Numeric Smooth Variable (ggplot)
}
\keyword{internal}
