% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implied_residuals.r
\name{plot_implied_residuals}
\alias{plot_implied_residuals}
\alias{plot_implied_residuals.gam}
\title{Plot Implied Residuals for GAM Models}
\usage{
plot_implied_residuals(
  gam_model,
  data,
  focus_var,
  var,
  nbins = 6,
  islog = FALSE,
  ylim = NULL,
  n.exclude = 0,
  ...
)

\method{plot_implied_residuals}{gam}(
  gam_model,
  data,
  focus_var,
  var,
  nbins = 6,
  islog = FALSE,
  ylim = NULL,
  n.exclude = 0,
  ...
)
}
\arguments{
\item{gam_model}{A fitted GAM model object (e.g., from `mgcv::gam`).}

\item{data}{The data frame used to fit the model.}

\item{focus_var}{The name of the focus variable (x-axis).}

\item{var}{The name of the variable for which to compute mean and quantiles (y-axis).}

\item{nbins}{(Optional) If `x_var` is not a factor, the number of bins to create.}

\item{islog}{Logical indicating if the y variable should be log-transformed.}

\item{ylim}{(Optional) The y-axis limits for the plot.}

\item{n.exclude}{(Optional) The minimum number of records required to include a group in the plot.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints a ggplot object (multi-panel plot), and (optionally) returns a list containing the plot and the data used for plotting.
}
\description{
This function generates a multi-panel plot to visualize the relationship between
predicted values of the GAM model's focus variable and the mean and quantiles of
another variable in the data.  It is a generic function.
}
