% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influence.R
\name{plot_influence}
\alias{plot_influence}
\title{Influence Plot for GAM Terms}
\usage{
plot_influence(x, panels = TRUE)
}
\arguments{
\item{x}{A gam_influence object with calculations performed}

\item{panels}{Whether to use separate panels for each term}
}
\value{
A ggplot object
}
\description{
Plot the influence of smooth terms on the parametric focus term
}
