% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_re.R
\name{plot_re}
\alias{plot_re}
\title{Random Effects Diagnostics}
\usage{
plot_re(obj, term, re_type = "qq")
}
\arguments{
\item{obj}{A `gam_influence` object.}

\item{term}{Random effect term name or numeric index of the term in the model.}

\item{re_type}{Type of diagnostic: "points", "qq", "hist", or "caterpillar". Default is "qq".}
}
\value{
A ggplot object showing the requested random effect diagnostic plot.
}
\description{
Plot diagnostics for random effects in a gam_influence object.
}
