% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_influence.R
\name{plot_step_influence}
\alias{plot_step_influence}
\title{Plot Step and Influence Together (ggplot)}
\usage{
plot_step_influence(x, panels = FALSE, ...)
}
\arguments{
\item{x}{gam_influence object}

\item{panels}{Logical, whether to use panel layout}

\item{...}{Additional arguments}
}
\value{
Combined ggplot object
}
\description{
Side-by-side step and influence plots
}
