% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_panels.R
\name{plot_step_panels}
\alias{plot_step_panels}
\title{Step plot for GAM Terms}
\usage{
plot_step_panels(step_effects, x, show_previous = TRUE, ncol = 1)
}
\arguments{
\item{step_effects}{A data frame containing the step effects}

\item{x}{A gam_influence object with calculations performed}

\item{show_previous}{Logical indicating whether to show previous steps}

\item{ncol}{(integer) Number of columns for faceting}
}
\value{
A ggplot object
}
\description{
Plot the step change of terms on the parametric focus term
}
\keyword{internal}
