% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_single.R
\name{plot_step_single}
\alias{plot_step_single}
\title{Step plot for GAM Terms}
\usage{
plot_step_single(step_effects, x, show_previous = TRUE)
}
\arguments{
\item{step_effects}{A data frame containing the step effects}

\item{x}{A gam_influence object with calculations performed}

\item{show_previous}{Logical indicating whether to show previous steps}
}
\value{
A ggplot object
}
\description{
Plot the step change of terms on the parametric focus term
}
\keyword{internal}
