% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_distribution.r
\name{plot_term_distribution}
\alias{plot_term_distribution}
\title{Plot Data Distribution for a Model Term}
\usage{
plot_term_distribution(obj, term)
}
\arguments{
\item{obj}{A `gam_influence` object.}

\item{term}{The character name of the model term to plot (e.g., "year" or "s(temp)").}
}
\value{
A ggplot object showing the data distribution.
}
\description{
Plots the data distribution for a specified model term and focus variable.
}
