% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_effects_data.R
\name{prepare_effects_data}
\alias{prepare_effects_data}
\title{Prepare Effects Data for ggplot}
\usage{
prepare_effects_data(x, terms)
}
\arguments{
\item{x}{gam_influence object}

\item{terms}{Terms to include}
}
\value{
Data frame for ggplot
}
\description{
Prepare Effects Data for ggplot
}
\keyword{internal}
