% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_step_data.R
\name{prepare_step_data}
\alias{prepare_step_data}
\title{Print Method for gam_influence Objects}
\usage{
prepare_step_data(step_effects, x, show_previous = TRUE)
}
\arguments{
\item{step_effects}{A list of step effects}

\item{x}{A gam_influence object}

\item{show_previous}{Logical indicating whether to show previous steps}
}
\description{
Print Method for gam_influence Objects
}
