% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma_family.r
\name{print.extended.family}
\alias{print.extended.family}
\title{Print Method for Extended Family Objects}
\usage{
\method{print}{extended.family}(x, ...)
}
\arguments{
\item{x}{An extended family object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
S3 print method for extended family objects created by gengamma()
}
