% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{print.gam_box_cox}
\alias{print.gam_box_cox}
\title{Print Method for Box-Cox Transformed GAM}
\usage{
\method{print}{gam_box_cox}(x, ...)
}
\arguments{
\item{x}{An object of class "gam_box_cox".}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for a Box-Cox transformed GAM.
}
