% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gam_influence.R
\name{print.gam_influence}
\alias{print.gam_influence}
\title{Print Method for Indices from gam_influence Object}
\usage{
\method{print}{gam_influence}(x, ...)
}
\arguments{
\item{x}{A `gam_influence` object that has been processed by `calculate_influence()`.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame containing the focus indices with descriptive column names.
#' @export
}
\description{
Extracts the focus indices from a `gam_influence` object, including
unstandardised and standardised indices, confidence intervals, and coefficients of variation (CV).
}
