% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gam_influence_combined.R
\name{print.gam_influence_combined}
\alias{print.gam_influence_combined}
\title{Print Method for Combined Influence Objects}
\usage{
\method{print}{gam_influence_combined}(x, ...)
}
\arguments{
\item{x}{A gam_influence_combined object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for objects of class 'gam_influence_combined'
}
