% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_influence_gam.R
\name{print.influence_gam}
\alias{print.influence_gam}
\title{Print influence_gam Object}
\usage{
\method{print}{influence_gam}(x, ...)
}
\arguments{
\item{x}{An object of class `influence_gam`.}

\item{...}{Additional arguments (unused).}
}
\description{
Print influence_gam Object
}
