% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_residual_patterns.R
\name{print.residual_pattern_analysis}
\alias{print.residual_pattern_analysis}
\title{Print Method for Residual Pattern Analysis}
\usage{
\method{print}{residual_pattern_analysis}(x, ...)
}
\arguments{
\item{x}{A residual_pattern_analysis object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for objects of class 'residual_pattern_analysis'
}
