% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_influence_summary.R
\name{print_influence_summary}
\alias{print_influence_summary}
\title{Print detailed influence summary}
\usage{
print_influence_summary(x)
}
\arguments{
\item{x}{gam_influence object}
}
\description{
Print detailed influence summary
}
