% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\alias{r2.influence_gam}
\title{r2}
\usage{
r2(obj, r2_type = c("r2", "r2Negel", "r2Dev"))

\method{r2}{influence_gam}(obj, r2_type = c("r2", "r2Negel", "r2Dev"))
}
\arguments{
\item{obj}{An object of class `influence_gam`.}

\item{r2_type}{the type of r2 statistic required.}
}
\value{
The `influence_gam` object with calculation results added.
}
\description{
Calcuate r2 summary statistics
}
\author{
Alistair Dunn
}
