% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_calculations.R
\name{reset_calculations}
\alias{reset_calculations}
\title{Reset Calculations}
\usage{
reset_calculations(x)
}
\arguments{
\item{x}{gam_influence object}
}
\value{
gam_influence object with calculations reset
}
\description{
Remove calculated results to allow recalculation with different parameters
}
