% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_influence_gam.R
\name{summary.influence_gam}
\alias{summary.influence_gam}
\title{Summarize influence_gam Object}
\usage{
\method{summary}{influence_gam}(object, ...)
}
\arguments{
\item{object}{An object of class `influence_gam`.}

\item{...}{Additional arguments (unused).}
}
\description{
Provides a summary, including the model comparison table and snippets
of indices and influences.
}
