% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package.R
\name{test_package}
\alias{test_package}
\title{Package Validation and Testing}
\usage{
test_package()
}
\value{
Logical indicating if tests pass
}
\description{
Test GAM Influence Package
}
\details{
Basic functionality test
}
