% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gam_influence.R
\name{validate_gam_influence}
\alias{validate_gam_influence}
\title{Validate GAM Influence Object}
\usage{
validate_gam_influence(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Logical indicating if valid
}
\description{
Validate GAM Influence Object
}
\keyword{internal}
