% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_influence.R
\name{calculate_influence}
\alias{calculate_influence}
\title{Perform Influence Calculations}
\usage{
calculate_influence(obj, ...)
}
\arguments{
\item{obj}{An object for which to perform calculations.}

\item{...}{Additional arguments passed to methods.}
}
\description{
A generic S3 function to perform calculations on a `gam_influence` object.
}
