% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_influence.influence_gam.r
\name{calculate_influence}
\alias{calculate_influence}
\alias{calculate_influence.influence_gam}
\title{calculate_influence}
\usage{
calculate_influence(obj, islog = NULL, verbose = TRUE, ...)

\method{calculate_influence}{influence_gam}(obj, islog = NULL, verbose = TRUE, ...)
}
\arguments{
\item{obj}{An object of class `influence_gam` created by `create_influence_gam`.}

\item{islog}{Deprecated. Logic now tries to infer from the influence object or the model.}

\item{verbose}{Logical. Print progress messages?}

\item{...}{Additional arguments (currently unused).}
}
\value{
The `influence_gam` object with calculation results added.
}
\description{
Performs the core calculations for influence analysis on a pre-created
`influence_gam` object. This populates the `indices`, `summary`,
`influences`, and `preds` slots of the object.
}
\author{
Alistair Dunn
}
