% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_influence.r
\name{gam_influence}
\alias{gam_influence}
\title{Create a GAM Influence Analysis Object}
\usage{
gam_influence(model, focus = NULL, data = NULL, response = NULL)
}
\arguments{
\item{model}{A fitted GAM model from mgcv (gam, bam, or gamm)}

\item{focus}{The focus term for influence analysis (smooth or parametric)}

\item{data}{Optional data frame used to fit the model}

\item{response}{Optional response variable name}
}
\value{
A 'gam_influence' S3 object
}
\description{
Creates an S3 object for analysing the influence of terms in a GAM model.
}
\examples{
\dontrun{
library(mgcv)
data(mcycle)
m1 <- gam(accel ~ s(times) + s(times, by = factor), data = mcycle)
gi <- gam_influence(m1, focus = "s(times)")
}
}
