% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_terms}
\alias{get_terms}
\title{Get terms from a gam_influence object}
\usage{
get_terms(x)
}
\arguments{
\item{x}{A gam_influence object}
}
\value{
Character vector of term names
}
\description{
Get terms from a gam_influence object
}
