% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdi.R
\name{plot_cdi}
\alias{plot_cdi}
\title{Coefficient-Distribution-Influence (CDI) Plot}
\usage{
plot_cdi(obj, term, re_type = "points")
}
\arguments{
\item{obj}{A `gam_influence` object.}

\item{term}{The character name of the model term to plot (e.g., `"s(temp)"`). Alternatively, it can be a numeric index of the term in the model. If a numeric index is provided, it will be converted to the corresponding term name.}

\item{re_type}{Character; for random effects, one of "points", "qq", "hist", or "caterpillar".}
}
\value{
A patchwork ggplot object.
}
\description{
Creates a Coefficient-Distribution-Influence (CDI) plot.
This is a multi-panel plot that visualises a term's effect, its data
distribution, and its influence on the focus term.
}
