% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implied_residuals.r
\name{plot_implied_residuals}
\alias{plot_implied_residuals}
\title{Implied Residuals Plot}
\usage{
plot_implied_residuals(
  obj,
  var,
  nbins = 6,
  islog = NULL,
  ylim = NULL,
  n.exclude = 0
)
}
\arguments{
\item{obj}{A `gam_influence` object containing calculated indices.}

\item{var}{The name of the variable for which to compute mean and quantiles (y-axis).}

\item{nbins}{(Optional) Number of bins if `var` is not a factor.}

\item{islog}{Logical indicating if the response variable should be log-transformed.}

\item{ylim}{(Optional) The y-axis limits for the plot.}

\item{n.exclude}{(Optional) Minimum number of records required to include a group in the plot.}
}
\value{
A ggplot object (multi-panel plot).
}
\description{
Creates a multi-panel plot showing the relationship between the focus term and another variable, including predicted values and observed means/quantiles.
}
