% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_re.r
\name{plot_re}
\alias{plot_re}
\title{Random Effects Diagnostics}
\usage{
plot_re(obj, term, re_type = "qq")
}
\arguments{
\item{obj}{A `gam_influence` object.}

\item{term}{Random effect term name.}

\item{re_type}{Type of diagnostic: "points", "qq", "hist", or "caterpillar"}
}
\description{
Plot diagnostics for random effects in a gam_influence object.
}
