% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_and_influence.R
\name{plot_step_and_influence}
\alias{plot_step_and_influence}
\title{Combined Step and Influence Plot}
\usage{
plot_step_and_influence(obj)
}
\arguments{
\item{obj}{A `gam_influence` object containing calculated indices.}
}
\value{
A patchwork plot combining the stepwise index and term influence plots.
}
\description{
Creates a combined plot showing both the stepwise index and term influence for a `gam_influence` object.
}
