% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gam_influence.R
\name{print.gam_influence}
\alias{print.gam_influence}
\title{Print Method for gam_influence Objects}
\usage{
\method{print}{gam_influence}(x, ...)
}
\arguments{
\item{x}{A gam_influence object}

\item{...}{Additional arguments}
}
\description{
Print Method for gam_influence Objects
}
