% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2.gam_influence}
\alias{r2.gam_influence}
\title{Model Progression Statistics for gam_influence}
\usage{
\method{r2}{gam_influence}(obj, ...)
}
\arguments{
\item{obj}{A `gam_influence` object that has been processed by `calculate_influence()`.}

\item{...}{Unused.}
}
\value{
A data frame containing the model progression statistics.
}
\description{
Method for extracting the model progression summary table from a calculated `gam_influence` object.
}
