% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Get Model Progression Statistics}
\usage{
r2(x)
}
\arguments{
\item{x}{A `gam_influence` object that has been processed by `calculate_influence()`.}
}
\value{
A data frame containing the model progression statistics.
}
\description{
Extracts the model progression summary table from a calculated `gam_influence` object.
This table includes statistics like AIC, R-squared, and deviance explained for each step of the model build.
}
