% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gam_influence.R
\name{summary.gam_influence}
\alias{summary.gam_influence}
\title{Summary Method for gam_influence Objects}
\usage{
\method{summary}{gam_influence}(object, ...)
}
\arguments{
\item{object}{A gam_influence object}

\item{...}{Additional arguments}
}
\description{
Summary Method for gam_influence Objects
}
